Special-case mode-name rewriting for fundamental mode
Emacs will not run hooks for «after-major-mode-change» for newly-created
fundamental mode buffers. In order to catch these, we now advise
«get-create-buffer».

An alternative would have been to hook into «buffer-list-update». However, this
is also emitted each time windows are selected, buffers killed, so that it is
run much more frequently. Additionally, the hook appears to only be run for the
buffer of the selected window, so that getting the newly created
fundamental-mode buffer would have been more involved.