Apply mode name rewriting to pre-existing buffers
Depending on where during Emacs initialization cyphejor-mode is engaged, it can
be the case that certain buffers we would like to have renamed already exist.
E.g., this will frequently be the case for the Messages buffer. In order to
remove the need for users to concern themselves with exactly when cyphejor-mode
is loaded, we now run «cyphejor--hook» for all buffers which are live when the
mode is activated.