Fix: don’t clobber match-data when running advice (#7)
When some command or mode is using match-data and our hook is triggered, e.g.,
from a call to ‘get-buffer-create’, Cyphejor would leave the match-data in a
modified state. This can confuse the caller of ‘get-buffer-create’ the next time
it runs ‘match-string’.

This issue can be reproduced by running ‘(shell-command "ls&")’, which will
raise an error.

We fix this by restoring match-data to its original state.